(PC-18372)[API] refactor: [WIP] split file api in folder api with different file inside
