+++ api/src/pcapi/core/educational/api/__init__.py
@@ -1,0 +1,8 @@
+from pcapi.core.educational.api.adage import *
+from pcapi.core.educational.api.booking import *
+from pcapi.core.educational.api.categories import *
+from pcapi.core.educational.api.deposit import *
+from pcapi.core.educational.api.institution import *
+from pcapi.core.educational.api.offer import *
+from pcapi.core.educational.api.stock import *
+from pcapi.core.educational.api.venue import *
